Fix memory leaking pages toolbar
Because of the particular way toolbars are constructed it makes them
incompatible with using GtkBuilder.

So to get around this we have to reference the widget so it's not
destroyed along with the builder itself. This reference was never
cleaned up, this fixes the leak by using a first-parent unref.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3249