Revert tolerance square calculation to fix smothing
Many users who use the pencil have complained that the smoothing has
become very aggressive. Looking into changes to the tolerance maths
since 0.92 points to the commit:

37a3e996eee803fb5d03a09037a71c79de77175d

Which increased the square of the tolerance by a factor of 50 and
because the tolerance is limited to 1-100 (zero is not allowed) it's
basically impossible to turn off aggressive smoothing in the pencil tool.

I can see no reason for this change for the LPE work or pencils in general.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3251