Fix an aliasing problem in ObjectSet::toLayer()
Moving a set of objects into a layer requires knowing a `pivot`
element after which the selection should be placed. An aliasing issue
occurs if the pivot is a part of the selection: when the selection is
cut out, the pivot gets destroyed and the paste operation crashes.
This commit introduces an aliasing check to avert these crashes.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3247
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3248