Fix a crash on erasing from a clipped and cloned object
This commit eliminates a crash which occurred when the "Clip" mode of
the Eraser tool was used on an object which had clip set and which
additionally was cloned. The crash was caused by the Eraser tool
deleting the old clipPath before setting the new one, which propagated
an invalid state to the clones.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1457