Fix page remove, add and selection issues
The removePage function was incorrectly returning when a page
was found to be in the middle of the pack, instead of inserting.

As well as fixing the main issue, cleaning up the insertion and
selection-on-delete removes warnings and makes the pages list
more robust.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3224
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3161