Re-Add the cut path and cut text deletion code
The deleteItems code was changed to only delete items
which makes a lot of sense. But this makes the cut path
and cut text delete the selected item instead of just
deleting the text or nodes.

This code path is put back in, but in the cut code directly.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3179