Improve consistency of Eraser tool's Cut mode
This commit changes the way the Eraser tool treats clones when in "Cut"
mode. Clones are unlinked only if the original cloned item is of a type
that can be successfully erased in the Cut mode.

Straight line segments and their clones are ignored by the "Cut" mode
regardless of how many intermediate nodes they contain. The same applies
to paths that combine several straight line segments in one object.
Erasing all other paths "just works", including those with 1 or 2 nodes.

Many thanks to Nathan Lee for suggesting and testing these improvements.