Detect existing grids with chess algorithm
In order to fix the bug, a new algorithm was required to detect where
items are in relation to other objects. The first problem the old
method has, is calculating on just two objects at a time, which is
a flawed method as you can't really tell if two objects are close
to the total row or just close to each other.

The new method calculates rows based on their proximity to a fixed
minimum y value. And calculates subsequent rows with a recursion on
the remaining items.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1700