Remember lost padding when grid arranging to bbox
Bounding box arrangement losses padding on the right and bottom sides
if the height and width are set to be equal for all rows and columns.

We add this padding back in so the output bbox is the same as the input.
Cleaned the code to remove function wide variables.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3314