Add browse for file loader to the welcome screen
The welcome screen was missing the ability to browse for a file
from the hard drive. User testing shows this to be a critical
ommission from the welcome screen.

Fixes an issue where it would quit if it failed to open a file.
Fixes the lack of version info in the title which is visible on windows.
Fixes closing dialog (Alt+F4) without loading new document.