Enable setting paint from the Paint Servers dialog
Selecting a shape and clicking on one of the hatches available in the
Paint Servers dialog will now set the paint to the selected hatch.

Unused paints are correctly cleaned up both from the <defs> and from the
list of hatches in the current document, which prevents a segfault that
could otherwise occur.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2711