Rework the loading and management of paint servers
Redesign the way in which the Paint Servers dialog manages the available
paints.

Refactor the process of loading available paint servers from the current
file and from the files in "share/paints". Deduplication of paints based
on their ids is now global, rather than per source document. The list of
stock paints no longer duplicates when switching documents.

The new deduplication mechanism fixes a crash which occurred when a copy
of a stock paint present in the current document was loaded into a newly
created Paint Servers dialog along with the original stock paint.

Redo the logic of how the list of all paint servers is regenerated when
the document is changed or replaced.

Fix a memory leak from `return new`.

Improve documentation of the class `PaintServersDialog`.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3327