Move tab navigation from actions to tools
The next and previous tab actions conflict with gtk's internal widget
navigation. There is no solution which involves keeping these keys
as actions, since we do not have event_context or canvas widget level
actions yet, so can not limit the accel bindings to when the canvas is
action.

This solution moves the functionality into the tool code, as the cost
of configurability. Select Next and Prev are no longer configurable keys.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3245