Canculate the canvas text extent correctly
The Cairo text extents function does not return the actual width and height
of the proposed text, instead it is a complex set of information which
depends on the direction the text will be printed in and the spaces around
the text.

This also fixes any issue with labels that has trailing or starting space.

The font needs to be set in rendering and bounds checkings to be effective.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3316