Remove function sp_document_namedview
This old function is not needed since the document object has it's own
namedview function that can check for namedview's existance and create
one if needed. This fixes crashes with files that have a namedview but
do not have it in the right place.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2843