Don't output hidden layers in PDF output
The picking of items on the page ignores layers which means it's already
looking at layer children which are not hidden, even if the layer is.

The picker has controls for this, so we extend the page item picker to
include the same functionality and make use of it in the pdf output.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3336