Select correct context for marker context-fill and stroke
The context object when populating the context-fill and context-stroke
was always set to be the first href the marker had seen in pdf export.

Instead we need to track the parent object and the clone context
so that we can make use of them. We also need to prevent poluting the
paint origin after using the marker's style structure.

One other fix included is an Gtk-CRITICAL in the unrefHref caused by
the order of the checks being the same as the refHref.wq

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2545