Prevent clumping of x/y/w/h changes in undo history
Consecutive adjustments of x position via the spinbutton in the Select
Tool's toolbar are only clumped into a single event in the undo history
if they apply to the same selection. Whenever the selection is changed,
the subsequent adjustments will constitute a new event in the history.
The same applies to y position, width and height.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3217