Fix a crash in grids when fitting page to drawing
This fixes a segfault that occurred with documents containing grids when
the user opened and then closed the Document Properties dialog.  In such
scenario, the CanvasGrid object held on to stale pointers to the widgets
present in the "Grids" tab of the dialog, resulting in a crash when the
grid was modified from outside the dialog, e.g. when resizing the page to
the drawing (Ctrl+Shift+R).

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3008