Apply stroke width adjustment inside Text Layout
This patch pulls together the various bits of the code that was adding the
stroke width of the local object. Moving it to the LayoutNG codebase allows
the inclusion of the stroke into the bounding box.

This doesn't include a better calculation, it's the same calculation as we
were doing before. But now it's more consistant and inline.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/869