Live update of patterns in Paint Servers dialog
Ensure that whenever a pattern in the <defs> of the document is added,
removed or modified, the "Current document" tab of the "Paint Servers"
dialog shows that change instantly.  This also removes an intermittent
glitch when a paint server is auto-vacuumed (issue #3356).

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3356
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3357