Calculate stroke from stroke only visual box
When transforming objects, we were calculating the size of the stroke from
the visual bounding box. This is fine most of the time, but causes a lot of
issues with blurs as a blur will look like a stroke to this transformation
code.

The solution is to only use the stroke visual box.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/725