Paint Servers dialog: highlight selection's paint
The preview tiles for available paint servers in the Paint Servers
dialog now react to changing the selection in the document. If the
selected objects share a common paint as their `fill`, the preview
of that paint will be highlighted in the icon view, assuming it is
present there. Otherwise, the highlight will be cleared. This also
happens for `stroke` if the radio buttons are set to target stroke.