Fix the creation of pages in a y-flipped desktop
The coordinates saved to the svg should not ever be flipped for pages
unlike the guidelines which are always flipped.

This change makes sure the document coordinates are what are saved and
loaded from the document.

It also fixes a bug in the canvas item rect shadow, which would y move
away from the other rectangle.

Adds four page resizing points to account for issues with node location.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3382