Remove spaces from marker ids
Currently the marker system is very broken, confusing stockid and
element id and when copying and pasting any marker with a space in
it's id, you get a broken url and the copied marker is deleted.

While fixing the css is the deeper fix, this should prevent the
error with shipped markers by removing the spaces.