Fix double opening of welcome screen
If the inkscape process is launched multiple times we want to
cancel to showing of any startup screen and load the needed
new document or specified file without keeping a window open.

We also add a close button to the top right corner for greater ux
and don't open the welcome screen if the user's first response was
canceling.