Fix incorrect handling of units in stroke style ui
This fixes a couple of different problems in the user interface for the
stroke unit selection. Firstly hairline couldn't be selected because when
changed, the strokeWidth still matched so it assumed no update was needed.

Hairline now sets a width of 1, instead of 0 in the widthSpin box.

Hairline left the widthSpin sensitive even though changing the value did
nothing. Now it is disabled when the selcted items as hairline. Selecting
Percent for a hairline object keeps the widget disabled b/c this is nonsense.

Secondly moving from percent to units and back again no longer clears the value
and it will bring back the value by asking for the line style to be refreshed.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3402